Move chacha tests
